feat: 로또 구입 기능 정의
# step 3

feat: 로또 구입 기능 정의

# step 3
