+++ src/main/kotlin/lotto/presentation/controller/PurchaseResponse.kt
@@ -1,0 +1,10 @@
+package lotto.presentation.controller
+
+import lotto.domain.LottoTickets
+
+class PurchaseResponse private constructor(val tickets: List<List<Int>>) {
+
+    companion object {
+        fun of(tickets: LottoTickets) = run { PurchaseResponse(tickets.tickets.map { it.numbers }) }
+    }
+}
